* implemented app routing using BrowserRouter and react-router-dom * refactored structure of /src to add organization for types and pages * implemented a 404 error page * refactored App.tsx content to Home.tsx * implemented template structure for pages with Lesson1.tsx
