+++ package.json
@@ -17,6 +17,7 @@
     "prettier": "^2.5.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.3.0",
     "stylelint-config-prettier": "^9.0.3"
   },
   "devDependencies": {
