+++ src/index.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react';
 import ReactDOM from 'react-dom';
 import './styles/global.scss';
-import App from './components/App';
+import App from './App';
 
 ReactDOM.render(
   <StrictMode>
