+++ src/pages/Lesson1.tsx
@@ -1,0 +1,15 @@
+import { FC } from 'react';
+
+const Lesson1: FC = () => {
+  return (
+    <>
+      <h1>Lesson 1</h1>
+      <p>
+        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer luctus
+        lacus at consequat rutrum.
+      </p>
+    </>
+  );
+};
+
+export default Lesson1;
