+++ src/types/globalTypes.ts
